% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_npoints.R
\name{TS_npoints}
\alias{TS_npoints}
\title{Batch Bayesian optimization using Thompson sampling}
\usage{
TS_npoints(model, npoints, Xgrid)
}
\arguments{
\item{model}{an object of class \code{hetGP}; e.g., as returned by \code{mleHetGP}}

\item{npoints}{an integer representing the desired number of samples}

\item{Xgrid}{a matrix of locations at which the samples are drawn}
}
\value{
a matrix containing the \code{npoints} best locations where next batch of simulations should be run
}
\description{
Batch Bayesian optimization using Thompson sampling
}
